migrate from Qt5 to Qt6
Linux developers will need to install the following packages:

apt install -y qt6-base-dev qt6-base-dev-tools libqt6opengl6-dev \
  libqt6websockets6-dev qtkeychain-qt6-dev

+++ src/mapstorage/mapstorage.cpp
@@ -668,7 +668,7 @@ bool MapStorage::virt_saveData(const MapLoadData &mapData)
                 .arg(QString::number(compressionRatio, 'f', 1)));
 
         // TODO: add progress counter
-        fileStream.writeRawData(compressedData.data(), compressedData.size());
+        fileStream.writeRawData(compressedData.data(), static_cast<int>(compressedData.size()));
     }
 
     log("Writing data finished.");
